Altered unique item generation
The original code always selected a unique item of a given UIItemId with the highest available UIMinLvl both when an item is spawned an when entering a new game.  Now, all uniques of a given UIItemId that meet the UIMinLvl requirements have an equal chance of being selected.  To prevent a unique item from morphing upon loading a game, when a unique item is generated, the UIMinLvl is stored in the item instance, rather than the mlvl.  Finally, unique items where UIItemId and UIMinLvl are the same have been changed so they have a unique UIMinLvl.  These changes should allow additional items unique items to be added and generated so long as all items with a given UIItemId have a unique UIMinLvl, and is less than the highest monster mLevel.