Fixed Diablo hit % bug
The mHit member of MonsterStruct was an unsigned char and was changed to an int.  On nightmare, monsters get +85 to hit on nightmare and +120 on hell.  Diablo's hit % on normal is 220, so the other difficulties caused it to overflow, giving him a really low hit %.  All the code that use it for calculations use ints, so this was an easy fix.